+++ appyx-components/internal/test-drive/android/src/androidTest/kotlin/com/bumble/appyx/components/internal/testdrive/android/KeyframesTest.kt
@@ -1,6 +1,5 @@
 package com.bumble.appyx.components.internal.testdrive.android
 
-import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
@@ -103,9 +102,6 @@ class KeyframesTest {
         composeTestRule.snapshot("${javaClass.simpleName}_${nameRule.methodName}")
     }
 
-    @OptIn(
-        ExperimentalMaterialApi::class
-    )
     @Test
     @Ignore("This test attempts to find flickers and artifacts that were fixed in the past. However the artifacts do not show up when running this test")
     fun validate_that_no_artefacts_appear_during_animation() {